+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Support for tuple-type arguments to `create()`.
 - Support for number and string fields in object-type arguments of `create()`.
 
+### Maintenance
+
+- Updated dev dependencies.
+
 ## v1.2.1 - 2021-05-18
 
 ### Maintenance