+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Maintenance
 
 - Updated dev dependencies.
+- Refactored benchmark scripts.
 
 ## v1.2.1 - 2021-05-18